refactor(scripts): move async job runner under src/scripts
